session-watcher: Add new module to react to data from sessions.
This module is a thin wrapper around a pthread that watches all existing
sessions as well as a special pipe that is used to signal the existance
of new sessions. The watcher is passed a reference to a session-manager
that is effectively a container for sessions that the watcher watches.
The watcher is not responsible for doing anything with new sessions
beyond watching them. When the watcher receives a signal indicating a
new session has been added to the session-manager all it does is start
watching it.

Signed-off-by: Philip Tricca <flihp@twobit.us>