session-watcher: Add callbacks for handling events.
I'm calling the two events that the session-watcher must respond to:
1) the wakeup event and 2) the session event. The first is used to
wakeup the watcher thread when a new session has been added to the
session-manager. The second is for handling data coming in from existing
sessions. By allowing the consumer to supply callbacks (complete with
user_data) for these events the session manager becomes significantly
more powerful / flexible and becomes much more testable to boot. See the
new unit test added to the session-watcher_unit test collection for an
example.

Signed-off-by: Philip Tricca <flihp@twobit.us>