session-watcher: Set fd_set on each call to select, handle closed sockets.
These were two big bugs in the loop that watches for incoming data from
clients. We now configure the fd_set before each call to select (like
you're supposed to do). Also when we're reading / writing data to / from
the pipes we handle the case where the client has closed the pipe. This
allows us to shutdown the server end of the connection and remove the
session from the session-manager properly.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>