tcti: Remove GMain thread.
This never should have been in the TCTI. The main thread belongs in the
application. When we replace all of the raw dbus stuff with gdbus we can
do away with the gmain thread and use only synchronous calls. Any
non-blocking calls will have to be setup to use the gmain thread from
the application.

Signed-off-by: Philip Tricca <flihp@twobit.us>