tss2-tabd: Use a separate thread to initialize stuff.
The goal of the main thread is to register dbus names as early as
possible and then get into the gmain thread. This allows us to become
responsive on the dbus as early as possible. Clients calling functions
that we need to respond to via the dbus will block until the init thread
finishes.

Signed-off-by: Philip Tricca <flihp@twobit.us>