CreateConnection: Generate random session ID, return to caller as out parameter.
This requires adding an 'id' field in the session-data structure for the
server to associte the id with a session. When we add additional
functions that require we be able to associate a dbus call passed a
session ID with the session object we'll need another hash table to do
the lookup.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>