session-watcher: Integrate TAB into session-watcher and daemon.
The session-watcher was a simple read / write echo up till now. With the
tab and response-watcher available now we update the session-watcher to
enqueue blobs (that presumably hold TPM2 commands) in the tab. The tab
does it's thing and when the responses come back the response-watcher
will pull them out of the tab, get the right FD for the session and then
send the response.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>