session-watcher: Don't free session_data_t after removing from session_manager_t
By using the free function in the g_hash_table we cause the
session_data_t objects to be free'd when they're removed. This means
that the hash tables have sole ownership of the session_data_t objects.
Well actually the session_from_id_table is the sole owner since it's the
one that invokes the free function.

This is yet another indication that we should be using the gobject
reference counting.

Signed-off-by: Philip Tricca <flihp@twobit.us>