DataMessage: Add new GObject to hold a size / blob combination.
This should eventually replace the blob_t, but that will require
replacing the blob_queue_t with a message queue. The unit test makes
sure we can allocate and deallocate DataMessage objects and that they
will free the blob of data that it owns. Valgrind ang GLib / GObject
don't quite get along so we need a suppression file for valgrind to
ignore expected behavior.

Signed-off-by: Philip Tricca <flihp@twobit.us>