Expose the critical TCTI functions through the interface object.
This prevents us from having to expose the TSS2_TCTI_CONTEXT pointer
to the consuming application. It also requires us to make a number of
changes to the tabd and Tab. In the main thread we now get the Tcti
object from the TctiOptions class. We defer the initialization of the
TCTI till the init thread is running.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>