Add functions to read TPM command buffers.
These functions are informed by the header structure in the TPM command
buffer. The header is fixed size and read before the variable length
body. We read the header first, then get the size of the body and read
that. Sanity checks for the size of the command body are included though
they're mostly arbitrary now. There may be an upper bound on the size of
a command buffer and if that's the case these should be updated
accordingly.

This includes a new module to get data out of the TPM command / response
headers.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>