Tab: Add functions to send the TPM Startup command when the Tab thread starts.
This is often times necessary for the simulator since there's no
firmware to do this for us like a real TPM. Executing this multiple
times even on a real TPM shouldn't cause problems according to the
specification. We detect both a successful execution of this command as
well as the RC that's returned if the TPM has already been Startup'd.
Any other RC will cause the program to halt (default g_error behavior).

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>