Move the output queue from the Tab to the ResonseWatcher.
This is part of a larger shift in the TPM command processing pipeline.
The core of this new design is that each stage (SessionWatcher, Tab and
ResponseWatcher) each must have an internal object to block a worker
thread on. A source (SessionWatcher in this case) has it's own internal
IO mechanism that it will be blocking on waiting for data so it doesn't
have a queue. The sink (ResponseWatcher) will not have a reference to
the previous element in the pipeline so the pipeline is effectively
created in reverse.

The added benefit of this is that each stage in the pipeline will block
its worker thread on an object (queue) that it owns. This makes it easy
to cancel threads since we must wake up the worker and we can do that
immediately in the cancel function.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>