SessionManager: Add 'new-session' signal.
This is a useful mechanism to encapsulate the ugly details about the way
the CommandSource internals work. Previously tabd module would create
the object for the SessionData, add it to the SessionManager and then
signal to the CommandSource via the 'wakeup' fds that a new SessionData
had been created. tabd was also responsible for creating the wakeup pipe
fds.

By adding a signal in the SessionManager that's triggered when a
SessionData instance is added to it (a new Session) we allow the
SessonManager to let the CommandSource know about the new SessionData
directly. In the CommandSource callback function it can do the signaling
to the CommandSource thread directly. This allows the CommandSource to
manage the 'wakeup' fds  internally so no one else (the tabd) needs to
know about them.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>