Sink: Add interface type for pipeline sinks and port the ResponseSink over.
The ResponseSink is a good first use for this since it's a sink but also
because it already implements the Thread interface so it's our first
multi-interface implementer.

As defined currently, each stage in the command processing pipeline that
isn't the source can implement the Sink interface (it's a single
'enqueue' function). I'm not sure it's worth the effort to have a
separate interface for Sink, Source and Filter. Likely we'll end up with
a Source/Sink interface and 'filter' will be defined as an object that
implements both the Source and Sink interfaces.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>