Add AccessBroker object with unit tests.
This object is intended to mediate access to the SAPI / TCTI for the
downstream TPM. There are two use cases we need to support. The first
is a simple Tpm2Command object being sent to the TPM. The other is
the more complex use would be locking the AccessBroker, using the
SAPI or TCTI context directly, then unlocking the access broker.

Likely the later use case will be the most common as the RM will be
executing more than one command as part of executing a single command
from the user.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>