Replace Tcti interface type with an abstract base type.
This requires fixups far and wide but it's a positive change. There was
an ugly bug in the interface code: I needed a place to stash the
TSS2_TCTI_CONTEXT pointer and so I put it in the class object for the
interface type. Normally you'd want to put this in an instance object
and the reason you would want to do this is that the class and instance
object lifecycle are very different. Freeing the context in the instance
finalize function will cause some really weird behavior with segfaults
but AFAIK only after creating and freeing more than one object instance.
I only ran across this because unit testing.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>