ResourceManager: Don't unref Tpm2Resposne. Don't check for NULL sink.
The Tpm2Response object produced will have just a single reference
outstanding. Passing it to the Sink doesn't change this since the Sink
interface doesn't take another referece (thought that may be the right
way to do this). So for now we leave the Sink alone and remove the call
to g_object_unref so the sink will inherit the reference.

This isn't the right place to check for a NULL sink pointer. This should
probably be moved to the thread start routine?

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>