CommandSource: Rewrite terrible unit test with a more targeted approach.
The test case re-written was an integration test, not a unit test. It's
now written to test a single function: command_source_session_responder.
This test exercises the 'success' flow and verifies success by mocking
the sink_enqueue function verifying that the Tpm2Command passed to the
function under test is the same one that we expect.

There's also some refactoring hiding in there, specifically the
tpm2_command_from_fd function was moved from the CommandSource module
into the Tpm2Command module where it belongs. It's effectively a
Tpm2Command constructor.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>