MessageQueue: Take reference to object on 'enqueue', dequeue inherits it.
My reference counting strategy here was pretty bad. The AccessBroker
shouldn't be unref'ing the Tpm2Command buffer after it is sent.
Similarly the ResourceManager shouldn't be ref'ing objects before
they're enqueued / passed to another thread.

Instead the MessageQueue that threads communicate through should take
a reference to the objects that it holds. This allows the object holding
the reference to unref it after enqueuing the object. The object
dequeuing the object from the MessageQueue then inherits the reference.
This means that the MessageQueue will not unref said object.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>