SessionData: Add HandleMap data member and expose with a 'get' function.
There's not much "design" in this: each session needs to maintain state
for "virtualizing" handles. The HandleMap holds this state. The
unfortunate part is that this turns the SessionData object into more of
a container. Might turn into future refactoring of SessionData to move
the connection data (fds etc) into a separate object. Something that
would be a peer to the HandleMap.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>