tcti-tabrmd: Detect EOF in the receive function, return NO_CONNECTION.
This was an error condition that wasn't being caught. The function
instead was returning success and a size of 0 which caused the SAPI code
to return a weird error (INSUFFICIENT_RESPONSE). This whole situation
resulted from the daemon screwing up the reference count on the
SessionData object for this connection which caused the connection to be
closed. This change gets us a "sane" error code and should save some
running around in circles in the future.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>