ResourceManager: Implement logic to "virtualize" handles for transient objects.
This commit brings together the HandleMap, HandleMapEntry and the recent
changes to the Tpm2Command/Response objects for accessing the handle
area of commands and responses.

The resource manager will now replace each physical handle for objects
loaded into the TPM in the response sent back to the caller with a
virtual handle. A virtual handle is just another handle in the same
range / type but allocated independently from the TPM. The mapping
from physical to virtual handles is maintained in the HandleMap with a
HandleMapEntry created by the ResourceManager.

We do not yet catch commands that destroy / deallocate handles in the
TPM but that will be the next logical step. Following soon after this is
saving / restoring TPM state between calls to isolate clients.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>