tabrmd: Add option to ignore loaded contexts (current default) or fail.
When initializing the resource manager it may be that the TPM is already
in use and a number of handles that we want to virtualize already
allocated. This patch adds a new command line option
'--fail-on-loaded-trans' which causes the tabrmd to fail initialization
if transient objects are already loaded in the TPM.

This requires a new data member in the options structure but it also
requires that the options structure be available to the initialization
thread. We move the options structure to the gmain_data_t structure to
work around this.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>