Tpm2Command: Add function to transform handles from virtual to physical.
This function wraps a rather complex mechanism in a nice way. All
necessary data is available to the Tpm2Command since it holds a
reference to it's own SessionData & HandleMap. Once this function is run
the ResourceManager can just process the handles from the command to do
whatever mgmt it needs to do.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>