HandleMapEntry: Add small function to set the physical handle value.
Each time an object context is flushed the handle is invalidated and a
new handle is assigned when the context is loaded. This is an extreme
simplifacation since an object can be saved but not flushed and if that
context is loaded and not flushed it could be that two instances of the
same object loaded. I'm avoiding these sort of situations with the
current resoruce management scheme.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>