SessionManager: Increment reference count when returning SessionData.
This is intended to prevent the SessionData object from being freed
while it's in use. My discipline with reference counting has been a bit
lax up till now. This also requires that we unref SessionData objects
now when we obtain them from these functions.

We also remove a comment block disabling a pile of unit tests for the
SessionManager. Not sure why these were disabled but they work.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>