configure: Disable valgrind sgcheck.
Benefiting from this tool would be nice but it's buggy. Running this on
the ResourceManager unit test produces an error and causes the test to
abort. The error message shows a pile of dwarf related warnings:

--19950-- warning: evaluate_Dwarf3_Expr: unhandled DW_OP_ 0x93

eventually an error is returned:

exp-sgcheck: sg_main.c:560 (add_blocks_to_StackTree): Assertion '!already_present' failed.

This error (according to the internet) is thrown when valgrind fails to
rewrite some part of the stack. The assertion makes it look like it's
trying to write data that already exists. This seems to be a bug related
to some interplay between valgrind and gcc. See:
https://root.cern.ch/phpBB3/viewtopic.php?t=21520

When we move to a new version of valgrind or gcc we should try to enable
this tool to see if the bug has been resolved.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>