ResourceManager: Refactor transient object management code.
This fixes a bad assumption from the first pass: Since we flush all
transient objects for each command processed there's no use in having
a hash table keyed to the physical handle. These handles are invalidated
as soon as a command completes. The previous logic would set the
physical handle to 0 when it was invalidated by a flush and so the
has table keyed to the physical handle was dropping entries on account
of the colision between keys (phandle set to 0).

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>