util: Refactor the process_control_message to process_control_code.
This refactoring was required to fix a memory leak in the handling of
ControlMessages that cause thread objects to end their thread. If the
call to pthread_testcancel is made before the reference count on the
ControlMessage object is decremented then the reference to the
ControlMessage is leaked. This was happening when the ResourceManager
and the ResponseSink were canceled.

The approach to fixing this was to no longer pass the message object
into the function instead only passing in the ControlCode. This way the
callers can decrement the reference count before calling
process_control_code and prevent the leak.

This complicates the code in the message processing loops of both the
ResourceManager and the ResponseSink but it's better than leaking an
object reference.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>