ResourceManager: Add failing unit test for bad second handle in Tpm2Command.
This test exercises the cmd_virt_to_phys function. The test passes the
function a Tpm2Command with the second handle in the handle area set to
an unknown virtual handle. This makes it impossible to translate said
handle to a physical one and so the function returns an RC set
appropriately (with the correct RC: TSS2_RESMGR_ERROR_LEVEL +
TPM_RC_HANDLE + TPM_RC_H + TPM_RC_2).

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>