ResourceManager: Add passing test case for context_saveflush.
This one is pretty straight forward. The function mostly just looks up a
physical handle from the provided entry, passes it to the AccessBroker
to do the real work, then updates the physical handle in the entry. We
mock away the AccessBroker context_saveflush and have it return success
for this test. The test is successful if the phandle in the entry is set
to 0 by the function under test.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>