ResourceManager: Add simple function to transform virtual handles in a Tpm2Command.
This function takes a Tpm2Command, a HandleMapEntry and an offset in the
handle area of the command. It extracts the context from the
HandleMapEntry, has the AccessBroker load the context which gets us back
a physical handle, sets this in the HandleMapEntry and then changes the
handle in the command object.

I've also added a unit test for a successful run of this command. The
ResourceManager unit test needs serious refactoring.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>