HandleMap: Set limit on entries in the map.
This adds a new parameter to the handle_map_new function that sets an
upper limit on the number of entries allowed in the map. We then use
this in the tabrmd and ResourceManager to enforce an upper bound on the
number of transient objects that each session can have loaded.

Lots of fixup to the unit tests given the change in the HandleMap API.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>