Fix implicit sign conversions.
I'd like to be able to add -Wconversion to the CFLAGS but the code
generated by the gdbus-codegen tool doesn't play well with this flag. It
may be possible to do some magic to disable this flag for that single
compilation unit but that will probably require compiling the code into a
separate utility library just as a way to allow filtering the flag out.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>