logging: Fix bug introduced by df589f2829dc.
Fixing a bug in the set_logger function introduced another. The default
case of "stdout" wasn't being handled in the prevous code and the
default simply caused a successful return value. The chagen from
df589f2829dc caused the default case to return an error instead.

This change handles the default 'stdout' case just like the 'syslog'
case now and 0 is returned. -1 is returned when an unkonwn logger is
encountered.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>