ResourceManager: Only flush_contexts that we're managing.
This fixes a bug in the resource_manager_flush_context function caused
by assuming all FlushContext commands contain transient object handles.
This is bad assumption since the caller may be trying to flush a session
handle.

We fix this by using the pattern we've established of extracting the
TPM_HT (handle type), switching on it, and only flushing contexts for
transient handles. We add an empty case for session handles to be filled
in when we implement management for sessions.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>