util: Refactor the command / response header & body read / write functions.
The code in the util module to read TPM commands and the code in the
TCTI module to read TPM responses was nearly identical. Implementing
unit tests for each would have been incredibly reduntant. By refactoring
these functions we can implemtn unit tests for the core commands as part
of testing the util module. The bits that are specific to the TCTI can
then be tested separately.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>