logging: Add macro for "critical" events.
g_error should only be used for "fatal" errors. That is, situations in
the program that should *never* occur. For this reason g_error causes a
core dump of the application (for debugging). I've been using g_error
for any situation where I wanted the daemon to print a message and exit.

g_critical gets us something between a warning and g_error but it
doesn't terminate the application. This macro just wraps g_critical and
a call to exit for these purposes.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>