Implement GetCapability for queries on loaded transient handles.
This requires changes across several objects:
HandleMap:      Add an additional function to export a list of keys.
Tpm2Command    : Add functions to allow the extraction of the
  capability, property and property count fields of a GetCapability
  command.
tpm2-header    : Add functions to allow setting header values.
ResourceManager: Add mechanism to collect the data needed for, and
  the response buffer to a GetCapability command querying the loaded
  transient object handles.

We also include an integration test that creates a configurable number
of transient objects and then iterates over the handles using the
GetCapability command.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>