Tpm2Response: Check buffer size is large enough before reading a handle.
It's possible for an error to be returned for a command that would
normally have a handle in the response. Since we trim down the allocated
response buffer in the AccessBroker we may end up trying to read a
handle from this response that doesn't have a buffer large enought to
hold a handle.

Adding this check ensures that we won't read beyond the end of the
response buffer looking for handles.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>