test: Session management functions & ordering.
The ResourceManager has a state machine to track session lifecycle as it
must respond to some commands differently depending on the state of a
session. These tests execute some variation of the StartAuthSession,
ContextSave, Load and Flush commands. Additionally, omitting the Flush
command is a good way to test the connection teardown logic. When the
client connection is closed / dies all loaded / saved sessions
associated with the connection must be flushed.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>