Clean up data read and client termination log messages.
The read_data message output a debug message when `read` returns -1.
This is an error condition. Using 'warning' makes more sense.

The CommandSource object largely doesn't care why the connection was
closed only that the associated data it be cleaned up properly. The
message output when a client connection fails shouldn't be a warning,
just a debug message.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>