Tpm2Command: Implement mechanism to expose command authorizations.
We expose the authorizations from the auth area of a command using a
'foreach' method much like glib collections. The caller provides a
callback function (with the GFunc signature) and an optional reference
to some data. This function is invoked for each authorization in the
auth area. The callback has two parameters: the first is a reference
to the start of the authorization, the second is the caller-provided
data. We also provide a few macros to make accessing data from the
authorization a bit less awkward.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>