ResourceManager: Implement session loading from auth area.
This uses the tpm2_command_foreach_auth function to iterate over all
authrizations in the command. For each auth, we look to see if the
handle is a sessoin handle and if so we load it.

We've added an additional parameter to the resource_manager_load_session
function such that we can now pass a flag indicating whether or not the
session will be flushed upon command completion. If the session will be
flushed (based on attributes in the auth area) then we simply remove it
from the structures that the RM uses to track sessions. This causes the
associated resources to be cleaned up and the TPM will take care of the
rest.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>