Add documentation about the TCTI client library.
This includes adding a section to README about the TCTI client library
as well as small overview man page. This man page is in section 7
(overview, conventions and miscellaneous).

We also clean up the existing man page for the tpm2-abrmd a bit removing
some nonsensical entries for 'ENVIRONMENT' and 'BUGS', as well as adding
a 'COLOPHON' section with a reference to the github page for the
project.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>