unit: Remove test to start / cancel / join the ResourceManager thread.
This function breaks on glib version 2.40 (the version packaged in
the travis-ci Ubuntu instance) when we run the unit tests under
valgrind. The GAsyncQueue barfs up a g_critical if the queue is destroyed
when a thread is blocked on it. This doesn't happen for glib 2.42 on my
test system (though it could just be a threading thing).

Either way this test isn't particularly useful in the context of the
ResourceManager object. It only serves to prove that the internal
'_start', '_cancel' and '_join' functions from the Thread object work as
advertised and that the internal thread in the ResposneSink gets
unblocked when '_cancel' is called.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>