tabrmd: Instantiate the TctiOptions object in main, not parse_opts.
The TctiOptions object was instantiated as a side effect of the
parse_opts function. It gets used in the main function immediately after
the call to parse_opts (where the TCTI is instantiated).

It's better form to instantiate this object in main. Then it's passed
into the parse_opts function where the TCTI options are parsed. Then we
use it as a factory to pump out a Tcti object based on the parsed
options.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>