CommandAttrs: Delay locking AccessBroker in _init_tpm till necessary.
Delaying this lock till after the call to access_broker_get_max_command
then the unlocking logic is a bit cleaner. It's also good to lock the
mutex as late as possible and it's not necessary for this function to
hold the lock when the call to _get_max_command is made.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>