travis: Run 'make check' and dump logs.
We're currently running 'make check-valgrind' which only returns an
error code if memcheck detects errors in the unit test. If the unit test
itself fails however memcheck will not indicate a failure. So we need to
run both 'make check' to detect failures in the unit tests, and then
'make check-valgrind' to get errors detected by memcheck.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>