build: Add distcheck target to travis, pass configure args to distcheck.
By running distcheck we get a full test of the VPATH build stuff, all of
the packaging checks and a run of 'make check' for all of our unit
tests. The only thing distcheck won't do for us (or at least I don't
know how to get it to do this) is to run 'check-valgrind'. So we still
have to do a build manually and run 'check-valgrind'.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>