travis.yml: Change 'language' to 'c', add clang, use sudo & script for install.
The cpp -> c thing was an oversight. We probably started with the
.travis.yml from the TSS and forgot to change this. It probably needs to
be changed in the TSS master branch as well.

Adding a script invoked by sudo is a work-around for some weird travis-ci
stuff. Travis installs the clang compiler under in a location that's not
on the root user's path. This makes 'sudo make install' fail for clang
on travis, but work just fine for gcc.

This is discussed on the travis-ci github issue page but the suggested
work around (use 'sudo -i') doesn't work any longer:
https://github.com/travis-ci/travis-ci/issues/3088

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>