build: Add _tabrmd.log to CLEANFILES.
This is specifically to fix an issue that I've only seen in travis
builds: When the 'distcheck' target is run by travis it fails
complaining that a file called './_tabrmd.log' was "left in the build
directory after distclean". I can't repro on my development system but
it's such a small and harmless change to add this to CLEANFILES that
I'm doing so just to get pass the error.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>