tcti-tabrmd: Replace use of glib type in _init_full with local enum.
Using GBusType directly seemed like the easiest thing to do at the time
but it causes a weird side effect for users: including the tcti-tabrmd.h
header requires that glib headers be on the include path. This is a bit
obnoxious.

This patch hides the GBusType by creating our own enum type to represent
the bus type.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>