tcti-tabrmd: Gracefully fail if D-Bus connection is not created
If creating a D-Bus connection fails, libtcti-tabrmd calls g_error() which
causes an abort() and the process to be terminated. This function should
not be used for expected errors, and failing to create a connection can
happen if for example the tpm2-abrmd.service was not started.

In this case, the program using the library could decide to fallback using
the device TCTI to access the TPM directly or just handle the error better
than being terminated due a SIGABRT signal.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>