logging: Use G_MESSAGES_DEBUG to enable debug & info logging for syslog handler.
This isn't "the right way" to do this but it's better than the current
state of affairs. This patch disables info and debug messages for the
syslog handler by default. If the G_MESSAGES_DEBUG environment variable
is set to the special value "all" then info and log messages are enabled.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>