IpcFrontend: Introduce new abstract class to abstract IPC frontend.
This abstract class is meant to be a common interface to IPC mechanisms
exposed to clients. It is meant to be subclassed for concrete
implementations so that the tabrmd / main module may expose a multiple
IPC mechanism to client applications.

This class has 'connect' and 'disconnect' methods that must be overridden
by subclasses to manage their IPC stuff. Since disconnection can be
triggered by external factors this class also implements a signal to
notify subscribers of disconnect events.

The provided unit test demonstrates how to inherit from the IpcFrontend
class and emit the signal.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>