INSTALL.md: Update dependency list and rewrite / organize sections.
The dependency list was missing a reference to the TSS. This has been
added and the dependencies for the test harness have been broken out
into their own list since they're not strictly required to build the
daemon.

Documentation for the creation of the tss user / account has been moved
to the front of the instructions.

Documentation for the configuration options have been integrated into
the section describing the configuration process. Previosly these were
left till the end of the document and this was causing some confusion.
They've also been clarified in a number of places with additional
detail.

This commit also adds an additional section for the 'post-install'
process. This includes instructions for causing the relevant daemons /
subsystems to reload configuration data which is required to run the
daemon as the tss user.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>