test: Integrate valgrind into int-log-compiler.sh
int-log-compiler.sh is instrumented to detect the autotools VALGRIND*
environment variables. If they're detected then valgrind is now started
for the tests private tabrmd instance as well as the test executable.

Due to the additional delay associated with the valgrind startup stuff
we add an additional environment variable (TABRMD_TEST_TCTI_RETRIES).
This variable  allows for a configurable number of retries when the
test harness tries to initialize the TCTI for a test. The default value
is 5. The int-log-compiler.sh sets this to 10.

Several tests are added to XFAIL_TESTS since valgrind now catches stuff
in the daemon.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>